Downgrade Mac OS image to get build to pass (#132)
Downgrade Mac OS image to macos-13 (which is x86) to get the build to pass. Many packages still do not have M1 (arm64+macos) wheels, and when trying to build from source some libraries are missing (e.g. HDF5 headers).

It's apparently possible to run brew install ... commands in the github actions to fix this, but I will leave this to future me.